app/requirements.txt
ultralytics
opencv-python-headless
pillow
albumentations
gradio
streamlit